+++ src/continuation/continuation_grouping.jl
@@ -5,14 +5,25 @@ import Mmap
 struct FeaturizeGroupAcrossParameter{A <: BasinMapFeaturizeGroup, E} <: GlobalContinuationAlgorithm
     bmap::A
     info_extraction::E
-    par_weight::Float64
 end
 
 """
     FeaturizeGroupAcrossParameter <: GlobalContinuationAlgorithm
-    FeaturizeGroupAcrossParameter(bmap::BasinMapFeaturizeGroup; kwargs...)
+    FeaturizeGroupAcrossParameter(bmap::BasinMapFeaturizeGroup [, info_extraction])
 
-A method for [`global_continuation`](@ref).
+A method for [`global_continuation`](@ref) that featurizes and groups
+trajectories across the whole parameter axis to establish a continuation of the groups.
+
+`info_extraction::Function` is a function that takes as an input a vector of feature-vectors
+(corresponding to a cluster) and returns a description of the cluster.
+By default, the centroid of the cluster is used as its description.
+This output becomes the `attractors` representation in [`GlobalContinuationOutput`](@ref).
+
+This method adds some information into the `other` field of `GlobalContinuationOutput`:
+- `"`
+
+## Description
+
 It uses the featurizing approach discussed in [`BasinMapFeaturizeGroup`](@ref)
 and hence requires an instance of that bmap as an input.
 When used in [`global_continuation`](@ref), features are extracted
@@ -24,22 +35,9 @@ to each parameter value they came from.
 
 This continuation method is based on, but strongly generalizes, the approaches
 in the papers [Gelbrecht2020](@cite) and [Stender2021](@cite).
-
-## Keyword arguments
-
-- `info_extraction::Function` a function that takes as an input a vector of feature-vectors
-  (corresponding to a cluster) and returns a description of the cluster.
-  By default, the centroid of the cluster is used.
-  This is what the `attractors_cont` contains in the return of `global_continuation`.
 """
-function FeaturizeGroupAcrossParameter(
-        bmap::BasinMapFeaturizeGroup;
-        info_extraction = mean_across_features,
-        par_weight = 0.0,
-    )
-    return FeaturizeGroupAcrossParameter(
-        bmap, info_extraction, par_weight
-    )
+function FeaturizeGroupAcrossParameter(bmap::BasinMapFeaturizeGroup)
+    return FeaturizeGroupAcrossParameter(bmap, mean_across_features)
 end
 
 function mean_across_features(fs)
@@ -54,26 +52,17 @@ function mean_across_features(fs)
 end
 
 function global_continuation(
-        continuation::FeaturizeGroupAcrossParameter, pcurve::Vector, ics::InitialConditionSampler;
+        continuation::FeaturizeGroupAcrossParameter, pcurve::Vector, sampler::InitialConditionSampler;
         show_progress = true,
     )
-    (; bmap, info_extraction, par_weight) = continuation
-    spp, n = length(ics), length(pcurve)
-    features = _get_features_pcurve(bmap, ics, n, spp, pcurve, show_progress)
-
-    # This is a special clause for implementing the MCBB algorithm (weighting
-    # also by parameter value, i.e., making the parameter value a feature)
-    # It calls a special `group_features` function that also incorporates the
-    # parameter value (see below). Otherwise, we call normal `group_features`.
-    # TODO: We have deprecated this special clause. In the next version we need to cleanup
-    # the source code and remove the `par_weight` and its special treatment in `group_features`.
-    if bmap.group_config isa GroupViaClustering && par_weight ≠ 0
-        labels = group_features(features, bmap.group_config; par_weight, plength = n, spp)
-    else
-        labels = group_features(features, bmap.group_config)
-    end
-    fractions_cont, attractors_cont =
-        label_fractions_across_parameter(labels, 1features, n, spp, info_extraction)
+    (; bmap, info_extraction) = continuation
+    # spp means 'samples per parameter'
+    spp, n = length(sampler), length(pcurve)
+    features = _get_features_pcurve(bmap, sampler, n, spp, pcurve, show_progress)
+    labels = group_features(features, bmap.group_config)
+    fractions_cont, attractors_cont = label_fractions_across_parameter(
+        labels, features, n, spp, info_extraction
+    )
     return fractions_cont, attractors_cont
 end